cmd/starcoin: add pub fn submit_txn for cli.state for common usage
